Unified mutation tests
