--- modules/doobie/src/test/scala/mutation/MutationSpec.scala
@@ -1,11 +1,0 @@
-// Copyright (c) 2016-2020 Association of Universities for Research in Astronomy, Inc. (AURA)
-// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
-
-package mutation
-
-import utils.DatabaseSuite
-import grackle.test.SqlMutationSpec
-
-final class MutationSpec extends DatabaseSuite with SqlMutationSpec {
-  lazy val mapping = MutationMapping.fromTransactor(xa)
-}
